+++ src/index.ts
@@ -91,7 +91,11 @@ function composeAccept(first: AcceptCallback | null, second: AcceptCallback): Ac
     return first ? (relpath) => first(relpath) && second(relpath) : second;
 }
 
-export function normalizeOptions(options: Options = {}) {
+export function normalizeOptions(options: Options | string = {}) {
+    if (typeof options === 'string') {
+        options = { basedir: options };
+    }
+
     const posix = Boolean(options.posix);
     const pathSep = posix ? path.posix.sep : path.sep;
     const basedir = path.resolve(options.basedir || process.cwd());
@@ -220,7 +224,7 @@ export function normalizeOptions(options: Options = {}) {
     };
 }
 
-export function scanFs(options: Options) {
+export function scanFs(options: Options | string) {
     async function collect(basedir: string, absdir: string, reldir: string): Promise<any> {
         const tasks = [];